Inclusion of language.txt files
The language.txt specifies the name and character encoding for the respective language.
Old "Language" enum is now a class.  This shouldn't break mods (unless they were calling an unimplemented method).

This change was made so the hard-coded list of languages can be removed in order to allow new languages to be added without code-side changes.