Update Recipes_FI.txt
Unclear whether the "two-way radio" on line 52 refers to walkie-talkies, HAM radios, or both. Opted to interpret it as a walkie-talkie, since line 45 is for HAM radios specifically; if this was incorrect, please @ me on on Discord with the context and I will fix it ASAP.

Also a bit unclear on the specific in-game context of lines 64-67 (get muffin/s, bacon bits, bacon rashers); same as with the other one, I can potentially improve the translation for these with more info.