ContextMenu_PTBR: Translate a single line that wasn't yet. Includes some orthography fixes.
Note: at line 355 the "Unfavorite" label is translated as "Desfavoritar" however this doesn't exists in the Brazilian Portuguese dictionary but exists in Portugal Portuguese one. I've left as is for now as I believe this is for some button and there isn't much room for a longer string if translated the other way.