1st revision of translation
Hello! I'm a certified English/Spanish translator specialized in videogames! I look forward to keep working on this file as there are some things that can be written differently to stay as close to the original as possible whilst mantaining the effect it produces on the player! I also noticed some jokes were omitted.

Please, note these translations/revisions are valid for all Spanish speakers all around the world.

I will keep working on this file and I look forward to see the changes implemented in my favourite game!

Thank you guys, hope you enjoy it!

(If this is not the correct procedure to post translations, please let me know as well!)

OR: ORIGINAL
TR: TRANSLATION
RV: REVISED
________________________________________
Authentic British cuisine! OR

¡Auténtica receta inglesa! TR

¡Auténtica cocina inglesa! RV

Explanation: wrong translation of the word "cuisine" as it is not "recipe" but "cocina".

Cocina, in this context, does not mean "kitchen".
__________________________________________________________

Try our new Biscuit and Gravy menu! OR

Prueba nuestra Galleta con Sirope! TR

¡Prueba nuestro nuevo menú, galleta con salsa! RV

Explanation: they are inviting us to try the new MENU which consists of "biscuit and gravy", not the new "Biscuit and Gravy". This is a severe mistake.

The correct translation for this is "galleta con salsa" per different cuisine articles consulted by me.

"Sirope" is "syrup" in Spanish which does not make sense as gravy is a "salsa". I changed that as well.

_

This. This is the message of cows. OR

Este es el mensaje de las vacas. TR

Este. Este es el mensaje de las vacas. RV

Explanation: fixed to look as the original, there's no need for us to avoid that repetition as the original and the translation have almost the same character length.

____

Feel the most nutritious taste explosion known to man. OR

Siente la más nutritiva explosión de sabor conocida. TR

Siente la explosión de sabor más nutritiva. RV

Explanation: wrong word order, passive constructions are not as common as in English. Opted for a different approach whilst keeping the meaning in mind.

____

I am here to tell you something mooo-tiful. OR

Estoy aquí para hablarles sobre algo delicioso. TR

Estoy aquí para contarles algo muuu-licioso. RV

Explanation: we can keep the joke in Spanish! It still makes sense and makes us laugh, why would we use "delicious"? Plus, the person speaking is not "there" to "talk" but to "tell" something. The word that is nearer to the original here would be "contar".

____

We sure will Tooks! For $49.99 who'd resist? OR

¡Claro Tooks! Por 49.99 dólares ¿quién puede resistirse? TR

¡Claro Tooks! Por 49.99 dólares ¿Quién podría resistirse? RV

Explanation: wrong tense for "would", also fixed the lack of mayusc inside the interrogation marks per Spanish language rules.

____

Ask your doctor today! OR

¡Pregunta a tu médico! TR

¡Preguntale a tu médico hoy! RV

Explanation: translator omitted "today".

___

It makes a difference to pick up your medication... OR

Hay una diferencia entre recoger tu medicación... TR

Hay una diferencia entre recoger su medicamento... RV

Explanation: per the "Diccionario Critico de Dudas Ingles-Español de Medicina" from translator Fernando A. Navarro, it is not recommended to use "medicación" as the word itself is a "calco léxico" from the English language.

It is preferred to use terms that are commonly used and do not have English roots like "medicamento" or "medicina". I chose to use "medicamento" since "medicina" can be mistaken for "medicine".

____

Welcome to my... robot laboratory. OR

Bienvenidos a mi laboratorio robot. TR

Bienvenidos a mi... laboratorio de robots. RV

Explanation: the "..." are not to be omitted as they are required to create a certain effect on the person hearing.